chatbot/requirements.txt
flask
python-dotenv
requests
gunicorn
google-generativeai
